requirements.txt
pandas==2.1.0
pyarrow==15.0.0
flake8==6.0.0
pre-commit==3.0.2
pytest==7.1.3
google-api-python-client==2.116.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
lxml==5.1.0
beautifulsoup4==4.12.3
gspread==6.0.2
